style: Rename pointing to inherited_ui.
And also merge values::*::pointing into values::*::ui.

Bug: 1460192
MozReview-Commit-ID: FM4gWEszahB
