+++ components/style/values/computed/ui.rs
@@ -4,4 +4,18 @@
 
 //! Computed values for UI properties
 
+use values::computed::Number;
+use values::computed::color::Color;
+use values::computed::url::ComputedImageUrl;
+use values::generics::ui as generics;
+
 pub use values::specified::ui::MozForceBrokenImageIcon;
+
+/// A computed value for the `caret-color` property.
+pub type CaretColor = generics::CaretColor<Color>;
+
+/// A computed value for the `cursor` property.
+pub type Cursor = generics::Cursor<CursorImage>;
+
+/// A computed value for item of `image cursors`.
+pub type CursorImage = generics::CursorImage<ComputedImageUrl, Number>;
