+++ components/style/values/generics/mod.rs
@@ -26,13 +26,13 @@ pub mod font;
 pub mod gecko;
 pub mod grid;
 pub mod image;
-pub mod pointing;
 pub mod position;
 pub mod rect;
 pub mod size;
 pub mod svg;
 pub mod text;
 pub mod transform;
+pub mod ui;
 pub mod url;
 
 // https://drafts.csswg.org/css-counter-styles/#typedef-symbols-type
